#1327 Cleanup Code
