+++ src/MoBi.Presentation/UICommand/RemoveMultipleModulesUICommand.cs
@@ -1,23 +1,22 @@
 using System.Collections.Generic;
-using MoBi.Presentation.Tasks;
+using MoBi.Presentation.Tasks.Interaction;
 using OSPSuite.Core.Domain;
-using OSPSuite.Core.Domain.Data;
 using OSPSuite.Presentation.UICommands;
 
 namespace MoBi.Presentation.UICommand
 {
    public class RemoveMultipleModulesUICommand : ObjectUICommand<IReadOnlyList<Module>>
    {
-      private readonly IObservedDataTask _observedDataTask;
+      private readonly IInteractionTasksForModule _interactionTasks;
 
-      public RemoveMultipleModulesUICommand(IObservedDataTask observedDataTask)
+      public RemoveMultipleModulesUICommand(IInteractionTasksForModule interactionTasks)
       {
-         _observedDataTask = observedDataTask;
+         _interactionTasks = interactionTasks;
       }
 
       protected override void PerformExecute()
       {
-         _observedDataTask.RemoveMultipleModules(Subject);
+         _interactionTasks.RemoveMultipleModules(Subject);
       }
    }
 }
