+++ tests/MoBi.Tests/Presentation/Tasks/InteractionTasksForModuleSpecs.cs
@@ -21,8 +21,9 @@ namespace MoBi.Presentation.Tasks
 
       protected override void Context()
       {
+         var dialogCreator = A.Fake<IDialogCreator>();
          _context = A.Fake<IInteractionTaskContext>();
-         sut = new InteractionTasksForModule(_context, new EditTaskForModule(_context));
+         sut = new InteractionTasksForModule(_context, new EditTaskForModule(_context), dialogCreator);
       }
    }
